refactor: Split implementations into separate modules
